move `wrap_group_into_range()` to `keymap.rs`
